refactor: Sections / 함수 위치 변경, dto 위치 변경
