+++ src/test/java/subway/domain/SectionsTest.java
@@ -9,6 +9,7 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.dto.SectionRemovalResult;
 
 class SectionsTest {
